requirements.txt
fastapi==0.115.6
uvicorn
torch==2.5.1
torchvision==0.20.1
Pillow==9.5.0